[#172419733] Moved all data functions into data.py
